backend/requirements.txt
asgiref==3.2.10
autopep8==1.5.7
backcall==0.2.0
backports.entry-points-selectable==1.1.0
black==21.9b0
certifi==2021.5.30
cffi==1.14.6
cfgv==3.3.1
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.3
Django==3.1
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-extra-fields==3.1.1
filelock==3.0.12
gunicorn==20.1.0
identify==2.2.15
idna==3.2
ipython==7.27.0
isort==5.9.3
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nodeenv==1.6.0
oauthlib==3.1.1
parso==0.8.2
pathspec==0.9.0
pickleshare==0.7.5
Pillow==8.3.1
platformdirs==2.4.0
pre-commit==2.15.0
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.10.0
PyJWT==2.1.0
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.9.30
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.1
traitlets==5.1.0
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.6
virtualenv==20.8.1
wcwidth==0.2.5
